+++ src/components/Legend.js
@@ -281,10 +281,10 @@ export default class Legend extends React.Component {
         if (this.props.style) {
             if (this.props.style instanceof Styler) {
                 style = this.props.style.legendStyle(category.key, type);
-            } else if (_.isObject(this.props.style)) {
-                style = this.props.style[category.key];
             } else if (_.isFunction(this.props.style)) {
                 style = this.props.style(category.key);
+            } else if (_.isObject(this.props.style)) {
+                style = this.props.style ? this.props.style[category.key] : defaultStyle;
             }
         }
         return style;
@@ -324,7 +324,7 @@ export default class Legend extends React.Component {
         return merge(
             true,
             defaultStyle[styleMode],
-            styleMap.symbol[styleMode] ? styleMap.symbol[styleMode] : {}
+            styleMap.symbol ? styleMap.symbol[styleMode] : {}
         );
     }
 
@@ -334,7 +334,7 @@ export default class Legend extends React.Component {
         return merge(
             true,
             defaultStyle[styleMode],
-            styleMap.label[styleMode] ? styleMap.label[styleMode] : {}
+            styleMap.label ? styleMap.label[styleMode] : {}
         );
     }
 
@@ -344,7 +344,7 @@ export default class Legend extends React.Component {
         return merge(
             true,
             defaultStyle[styleMode],
-            styleMap.value[styleMode] ? styleMap.value[styleMode] : {}
+            styleMap.value ? styleMap.value[styleMode] : {}
         );
     }
 
@@ -378,13 +378,21 @@ export default class Legend extends React.Component {
 
         if (this.props.stack) {
             return (
-                <Flexbox justifyContent={align} flexDirection={"column"} marginBottom={"20px"}>
+                <Flexbox
+                    justifyContent={align}
+                    flexDirection={"column"}
+                    marginBottom={this.props.marginBottom}
+                >
                     {items}
                 </Flexbox>
             );
         } else {
             return (
-                <Flexbox justifyContent={align} flexWrap={"wrap"} marginBottom={"20px"}>
+                <Flexbox
+                    justifyContent={align}
+                    flexWrap={"wrap"}
+                    marginBottom={this.props.marginBottom}
+                >
                     {items}
                 </Flexbox>
             );
@@ -403,8 +411,8 @@ Legend.propTypes = {
      * Alignment of the legend within the available space. Either left or right.
      */
     align: PropTypes.oneOf(["left", "right"]),
-    style: PropTypes.oneOfType([PropTypes.object, PropTypes.func, PropTypes.instanceOf(Styler)])
-        .isRequired,
+
+    style: PropTypes.oneOfType([PropTypes.object, PropTypes.func, PropTypes.instanceOf(Styler)]),
 
     /**
      * The categories array specifies details and style for each item in the legend. For each item:
@@ -469,7 +477,13 @@ Legend.propTypes = {
     /**
      * Defines whether to stack legend items vertically or not
      */
-    stack: PropTypes.bool
+    stack: PropTypes.bool,
+
+    /**
+     * The margin at the bottom passed to the FlexBox component
+     * Default value is 20px
+     */
+    marginBottom: PropTypes.string
 };
 
 Legend.defaultProps = {
@@ -479,5 +493,6 @@ Legend.defaultProps = {
     align: "left",
     symbolWidth: 16,
     symbolHeight: 16,
-    stack: false
+    stack: false,
+    marginBottom: "20px"
 };